+++ src/components/spa/refine.test.tsx
@@ -1,6 +1,8 @@
 import { renderWithAccordion } from "utils/test-utils";
 import { screen, waitFor } from "@testing-library/react";
 import { RefinementStep } from "./refine";
+import { http, HttpResponse } from "msw";
+import { server } from "mocks/server";
 
 describe("SPA Refinement Step", () => {
   it("should render B-factor plot", async () => {
@@ -13,5 +15,43 @@ describe("SPA Refinement Step", () => {
     renderWithAccordion(<RefinementStep autoProcId={1} />);
 
     await waitFor(() => expect(screen.getByText("0.200")).toBeInTheDocument());
+  });
+
+  it("should render up to two 3D visualisations", async () => {
+    renderWithAccordion(<RefinementStep autoProcId={1} />);
+
+    await screen.findByText("Open 3D Visualisation (C8)");
+    expect(screen.getByText("Open 3D Visualisation (C1)")).toBeInTheDocument();
+  });
+
+  it("should only render one button if there's only one particle classification group", async () => {
+    server.use(
+      http.get(
+        "http://localhost/autoProc/:procId/classification",
+        () =>
+          HttpResponse.json({
+            items: [
+              {
+                particleClassificationId: 1,
+                classDistribution: 999,
+                classNumber: 1,
+                particlesPerClass: 1,
+                batchNumber: 155,
+                symmetry: "C1",
+                type: "2D",
+                selected: false,
+                bFactorFitIntercept: 10,
+                bFactorFitLinear: 10,
+              },
+            ],
+          }),
+        { once: true }
+      )
+    );
+
+    renderWithAccordion(<RefinementStep autoProcId={1} />);
+
+    await screen.findByText("Open 3D Visualisation (C1)");
+    expect(screen.getAllByText(/open 3d visualisation/i)).toHaveLength(1);
   });
 });